requirements.txt
streamlit>=1.32.0
polars>=0.20.0
duckdb>=0.10.0
rapidfuzz>=3.6.0
openai>=1.12.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
